fix: address PR review comments and failing CI
Stop treating every HTTP 410 as a session expiry, keep verified-but-unsafe
workflow findings in the report, and tighten path allowlisting. Fix the
dosage-calendar clamp type error, password-change remount, and analyzer
warnings that failed Frontend CI. Install Chrome for Mermaid CLI in docs CI.

+++ frontend/lib/features/patient/patient_dosage_calendar_page.dart
@@ -39,7 +39,7 @@ class _PatientDosageCalendarPageState extends State<PatientDosageCalendarPage> {
   void _loadMoreData() {
     if (_loadedMonths < _maxLoadedMonths) {
       setState(() {
-        _loadedMonths = (_loadedMonths + 3).clamp(1, _maxLoadedMonths);
+        _loadedMonths = (_loadedMonths + 3).clamp(1, _maxLoadedMonths).toInt();
       });
     }
   }
